Removes un-used attributes and some naming cleanup
