+++ tools/misbehaving-jmx-server/src/main/java/org/datadog/misbehavingjmxserver/FourAttributeMetricMBean.java
@@ -1,0 +1,11 @@
+package org.datadog.misbehavingjmxserver;
+
+public interface FourAttributeMetricMBean {
+    Number getNumberValue();
+
+    Double getDoubleValue();
+
+    Float getFloatValue();
+
+    Boolean getBooleanValue();
+}
