--- tools/misbehaving-jmx-server/src/main/java/org/datadog/misbehavingjmxserver/MetricMBean.java
@@ -1,21 +1,0 @@
-package org.datadog.misbehavingjmxserver;
-
-import javax.management.openmbean.TabularData;
-import javax.management.openmbean.TabularDataSupport;
-
-public interface MetricMBean {
-    String getName();
-
-    Number getNumberValue();
-
-    Double getDoubleValue();
-
-    Float getFloatValue();
-
-    Boolean getBooleanValue();
-
-    TabularData getTabularData();
-
-    TabularDataSupport getTabularDataSupport();
-
-}
