--- tools/misbehaving-jmx-server/src/main/java/org/datadog/misbehavingjmxserver/MyInt.java
@@ -1,17 +1,0 @@
-package org.datadog.misbehavingjmxserver;
-
-public class MyInt implements MyIntMBean {
-    private int counter;
-
-    public MyInt() {
-    }
-
-    public synchronized int getCounter() {
-        return counter;
-    }
-
-    public synchronized void setCounter(int value) {
-        counter = value;
-    }
-}
-
